requirements.txt
anthropic>=0.25.0
feedparser>=6.0.11
requests>=2.31.0
python-dotenv>=1.0.1
schedule>=1.2.2
jinja2>=3.1.4
sendgrid>=6.11.0
